requirements.txt
djangorestframework==3.16.1
djongo==1.3.7
pymongo==3.11.4
pillow==11.3.0
